+++ ebpf/headers/common.h
@@ -117,40 +117,19 @@ enum {
     }                                                 \
 })
 
-// for x86_64
-struct pt_regs {
-	long unsigned int r15;
-	long unsigned int r14;
-	long unsigned int r13;
-	long unsigned int r12;
-	long unsigned int bp;
-	long unsigned int bx;
-	long unsigned int r11;
-	long unsigned int r10;
-	long unsigned int r9;
-	long unsigned int r8;
-	long unsigned int ax;
-	long unsigned int cx;
-	long unsigned int dx;
-	long unsigned int si;
-	long unsigned int di;
-	long unsigned int orig_ax;
-	long unsigned int ip;
-	long unsigned int cs;
-	long unsigned int flags;
-	long unsigned int sp;
-	long unsigned int ss;
-};
-
-
-// TODO: define in compiler flags, compiler defines it by default
-#ifndef __TARGET_ARCH_x86
-	#define __TARGET_ARCH_x86
-	#define bpf_target_x86
-	#define bpf_target_defined
-#endif
-
-
 #ifndef __VMLINUX_H__
 #define __VMLINUX_H__
-#endif /* bpf_target_x86 */
+
+// #if defined(__TARGET_ARCH_x86)
+//     #define bpf_target_x86
+//     #define bpf_target_defined
+// #elif defined(__TARGET_ARCH_arm64)
+//     #define bpf_target_arm64
+//     #define bpf_target_defined
+// #else
+//     #undef bpf_target_defined
+// #endif
+#endif /* __VMLINUX_H__ */
+
+
+
